Refactor element extraction code into a separate source file. Trying to share this with the readonly backend.
